+++ src/stateMachine/mutation/BQL/enrich.ts
@@ -4,16 +4,16 @@ import type { TraversalCallbackContext } from 'object-traversal';
 import { traverse } from 'object-traversal';
 import { isArray, isObject } from 'radash';
 import { getCurrentFields, getCurrentSchema, getFieldSchema } from '../../../helpers';
-import type { BQLMutationBlock, EnrichedBQLMutationBlock, EnrichedBormSchema } from '../../../types';
+import type { BQLMutationBlock, BormConfig, EnrichedBQLMutationBlock, EnrichedBormSchema } from '../../../types';
 import { replaceToObj } from './enrichSteps/replaces';
 import { setRootMeta } from './enrichSteps/rootMeta';
 import { splitMultipleIds } from './enrichSteps/splitIds';
 import { enrichChildren } from './enrichSteps/enrichChildren';
 import { computeFields } from './enrichSteps/computeFields';
-import { unlinkAll } from './enrichSteps/unlinkAll';
 import { preHookValidations } from './enrichSteps/preHookValidations';
 import { preHookTransformations } from './enrichSteps/preHookTransformations';
 import { doAction } from './shared/doActions';
+import { unlinkAll } from './enrichSteps/unlinkAll';
 
 /*
 const getParentBzId = (node: BQLMutationBlock) => {
@@ -58,6 +58,7 @@ const dataFieldStep = (node: BQLMutationBlock, field: string) => {
 export const enrichBQLMutation = (
 	blocks: BQLMutationBlock | BQLMutationBlock[] | EnrichedBQLMutationBlock | EnrichedBQLMutationBlock[],
 	schema: EnrichedBormSchema,
+	config: BormConfig,
 ): EnrichedBQLMutationBlock | EnrichedBQLMutationBlock[] => {
 	//console.log('Before enrich', JSON.stringify(blocks, null, 2));
 
@@ -113,7 +114,11 @@ export const enrichBQLMutation = (
 						/// 3.2.1 replaces
 						//console.log('Before replace', JSON.stringify(isDraft(node) ? current(node) : node, null, 2));
 						if (['linkField', 'roleField'].includes(fieldSchema.fieldType)) {
-							replaceToObj(node, field, fieldSchema);
+							if (node[field] === null) {
+								unlinkAll(node, field, fieldSchema);
+							} else {
+								replaceToObj(node, field, fieldSchema);
+							}
 						}
 						//console.log('After replace', JSON.stringify(isDraft(node) ? current(node) : node, null, 2));
 
@@ -164,11 +169,7 @@ export const enrichBQLMutation = (
 						/// 3.2.4 children enrichment
 						//redefining childrenArray as it might have changed
 						if (['linkField', 'roleField'].includes(fieldSchema.fieldType)) {
-							if (node[field] === null) {
-								unlinkAll(node, field, fieldSchema);
-							} else {
-								enrichChildren(node, field, fieldSchema, schema);
-							}
+							enrichChildren(node, field, fieldSchema, schema);
 						}
 
 						//3.2.5 splitIds()
@@ -227,12 +228,12 @@ export const enrichBQLMutation = (
 
 						// 3.3.8
 						//#region pre-hook transformations
-						preHookTransformations(node, field, schema);
+						preHookTransformations(node, field, schema, config);
 						//#endregion pre-hook transformations
 
 						// 3.2.9
 						//#region pre-hook validations
-						preHookValidations(node, field, schema);
+						preHookValidations(node, field, schema, config);
 						//#endregion pre-hook validations
 					}
 				});